[JAVA-15031] Upgraded code for apache httpclient  + moved old version to apache-httpclient4 module.
